backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
pymysql==1.1.1
pydantic-settings==2.6.1
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
joblib==1.4.2
shap==0.46.0
